+++ framework/console/controllers/CacheController.php
@@ -271,6 +271,12 @@ class CacheController extends Controller
                 $caches[$name] = $component['class'];
             } elseif (is_string($component) && $this->isCacheClass($component)) {
                 $caches[$name] = $component;
+            } elseif ($component instanceof \Closure) {
+                $cache = Yii::$app->get($name);
+                if ($this->isCacheClass($cache)) {
+                    $cacheClass = get_class($cache);
+                    $caches[$name] = $cacheClass;
+                }
             }
         }